requirements.txt
flask
google-genai
pydantic
mcp
web3